ARTEMIS-5444 refactor temp-queue-namespace
The `temporary-queue-namespace` attribute was added via 6be8966 to deal
with the fact that when JMS clients create temporary destinations (e.g.
via `Session.createTemporaryQueue()` or `Session.createTemporaryTopic()`
) a UUID is used for the queue and/or address which means the name
can't be matched via any `address-setting` other than `#`.

The problem with this solution is that it was predicated on the fact
that the resources were temporary. However, it turns out that this was
a red herring. In the Core broker any queue can be a temporary queue
and applying the settings from the `temporary-queue-namespace` to
temporary resources rather than the actual match (if it exists)
produces confusing and undesired results.

The real issue isn't that the resources are temporary, but rather that
they are named with a UUID. This commit changes the semantic so that
the broker detects the UUID name and applies the namespace based on
that. It also does the following:

 - Deprecates `temporary-queue-namespace`
 - Updates all relevant documentation
 - Simplifies a bunch of code since we don't need to pass around a
   boolean indicating whether the queue is temporary
 - Adds tests for detecting valid UUIDs
 - Implements proper support JMS temporary topics for AMQP, Core, and
   OpenWire (including proper clean-up)
 - Updates XML parser as well as XSD, etc. to support new parameter
 - Adds support for `security-settings` for UUIDs which is now much
   simpler than it would have been without this refactoring

UUID detection is off by default since the default value for
`uuid-namespace` is empty. I tried to ensure detection was as fast as
possible so when it is active it has a minimal impact.

+++ tests/integration-tests/src/test/java/org/apache/activemq/artemis/tests/integration/security/SecurityTest.java
@@ -74,6 +74,7 @@ import org.apache.activemq.artemis.spi.core.security.jaas.NoCacheLoginException;
 import org.apache.activemq.artemis.spi.core.security.jaas.UserPrincipal;
 import org.apache.activemq.artemis.tests.util.ActiveMQTestBase;
 import org.apache.activemq.artemis.utils.CompositeAddress;
+import org.apache.activemq.artemis.utils.RandomUtil;
 import org.apache.activemq.artemis.utils.SensitiveDataCodec;
 import org.apache.activemq.artemis.utils.Wait;
 import org.apache.activemq.command.ActiveMQQueue;
@@ -554,7 +555,16 @@ public class SecurityTest extends ActiveMQTestBase {
 
    @Test
    public void testJAASSecurityManagerAuthorizationNegative() throws Exception {
-      final SimpleString ADDRESS = SimpleString.of("address");
+      internalTestJAASSecurityManagerAuthorizationNegative(false);
+   }
+
+   @Test
+   public void testJAASSecurityManagerAuthorizationNegativeWithUuidNamespace() throws Exception {
+      internalTestJAASSecurityManagerAuthorizationNegative(true);
+   }
+
+   private void internalTestJAASSecurityManagerAuthorizationNegative(boolean useUuidNamespace) throws Exception {
+      final SimpleString ADDRESS = useUuidNamespace ? SimpleString.of(RandomUtil.randomUUIDString()) : SimpleString.of("address");
       final SimpleString DURABLE_QUEUE = SimpleString.of("durableQueue");
       final SimpleString NON_DURABLE_QUEUE = SimpleString.of("nonDurableQueue");
 
@@ -562,7 +572,13 @@ public class SecurityTest extends ActiveMQTestBase {
       ActiveMQServer server = addServer(ActiveMQServers.newActiveMQServer(createDefaultInVMConfig().setSecurityEnabled(true), ManagementFactory.getPlatformMBeanServer(), securityManager, false));
       Set<Role> roles = new HashSet<>();
       roles.add(new Role("programmers", false, false, false, false, false, false, false, false, false, false, false, false));
-      server.getConfiguration().putSecurityRoles("#", roles);
+      String match = "#";
+      if (useUuidNamespace) {
+         final SimpleString UUID_NAMESPACE = RandomUtil.randomUUIDSimpleString();
+         server.getConfiguration().setUuidNamespace(UUID_NAMESPACE.toString());
+         match = UUID_NAMESPACE + ".#";
+      }
+      server.getConfiguration().putSecurityRoles(match, roles);
       server.start();
       server.addAddressInfo(new AddressInfo(ADDRESS, RoutingType.ANYCAST));
       server.createQueue(QueueConfiguration.of(DURABLE_QUEUE).setAddress(ADDRESS).setRoutingType(RoutingType.ANYCAST));
@@ -572,12 +588,13 @@ public class SecurityTest extends ActiveMQTestBase {
       ClientSession session = addClientSession(cf.createSession("first", "secret", false, true, true, false, 0));
 
       // CREATE_DURABLE_QUEUE
-      try {
-         session.createQueue(QueueConfiguration.of(DURABLE_QUEUE).setAddress(ADDRESS));
-         fail("should throw exception here");
-      } catch (ActiveMQException e) {
-         assertTrue(e.getMessage().contains("User: first"));
-         assertTrue(e.getMessage().contains("does not have permission='CREATE_DURABLE_QUEUE' for queue durableQueue on address address"));
+      if (!useUuidNamespace) {
+         try {
+            session.createQueue(QueueConfiguration.of(DURABLE_QUEUE).setAddress(ADDRESS));
+            fail("should throw exception here");
+         } catch (ActiveMQException e) {
+            assertTrue(e.getMessage().contains("User: first does not have permission='CREATE_DURABLE_QUEUE' for queue durableQueue on address " + ADDRESS));
+         }
       }
 
       // DELETE_DURABLE_QUEUE
@@ -585,8 +602,7 @@ public class SecurityTest extends ActiveMQTestBase {
          session.deleteQueue(DURABLE_QUEUE);
          fail("should throw exception here");
       } catch (ActiveMQException e) {
-         assertTrue(e.getMessage().contains("User: first"));
-         assertTrue(e.getMessage().contains("does not have permission='DELETE_DURABLE_QUEUE' for queue durableQueue on address address"));
+         assertTrue(e.getMessage().contains("User: first does not have permission='DELETE_DURABLE_QUEUE' for queue durableQueue on address " + ADDRESS));
       }
 
       // CREATE_NON_DURABLE_QUEUE
@@ -594,8 +610,7 @@ public class SecurityTest extends ActiveMQTestBase {
          session.createQueue(QueueConfiguration.of(NON_DURABLE_QUEUE).setAddress(ADDRESS).setDurable(false));
          fail("should throw exception here");
       } catch (ActiveMQException e) {
-         assertTrue(e.getMessage().contains("User: first"));
-         assertTrue(e.getMessage().contains("does not have permission='CREATE_NON_DURABLE_QUEUE' for queue nonDurableQueue on address address"));
+         assertTrue(e.getMessage().contains("User: first does not have permission='CREATE_NON_DURABLE_QUEUE' for queue nonDurableQueue on address " + ADDRESS));
       }
 
       // DELETE_NON_DURABLE_QUEUE
@@ -603,8 +618,7 @@ public class SecurityTest extends ActiveMQTestBase {
          session.deleteQueue(NON_DURABLE_QUEUE);
          fail("should throw exception here");
       } catch (ActiveMQException e) {
-         assertTrue(e.getMessage().contains("User: first"));
-         assertTrue(e.getMessage().contains("does not have permission='DELETE_NON_DURABLE_QUEUE' for queue nonDurableQueue on address address"));
+         assertTrue(e.getMessage().contains("User: first does not have permission='DELETE_NON_DURABLE_QUEUE' for queue nonDurableQueue on address " + ADDRESS));
       }
 
       // PRODUCE
@@ -613,8 +627,7 @@ public class SecurityTest extends ActiveMQTestBase {
          producer.send(session.createMessage(true));
          fail("should throw exception here");
       } catch (ActiveMQException e) {
-         assertTrue(e.getMessage().contains("User: first"));
-         assertTrue(e.getMessage().contains("does not have permission='SEND' on address address"));
+         assertTrue(e.getMessage().contains("User: first does not have permission='SEND' on address " + ADDRESS));
       }
 
       // CONSUME
@@ -622,8 +635,7 @@ public class SecurityTest extends ActiveMQTestBase {
          ClientConsumer consumer = session.createConsumer(DURABLE_QUEUE);
          fail("should throw exception here");
       } catch (ActiveMQException e) {
-         assertTrue(e.getMessage().contains("User: first"));
-         assertTrue(e.getMessage().contains("does not have permission='CONSUME' for queue durableQueue on address address"));
+         assertTrue(e.getMessage().contains("User: first does not have permission='CONSUME' for queue durableQueue on address " + ADDRESS));
       }
 
       // MANAGE
@@ -632,8 +644,7 @@ public class SecurityTest extends ActiveMQTestBase {
          producer.send(session.createMessage(true));
          fail("should throw exception here");
       } catch (ActiveMQException e) {
-         assertTrue(e.getMessage().contains("User: first"));
-         assertTrue(e.getMessage().contains("does not have permission='MANAGE' on address activemq.management"));
+         assertTrue(e.getMessage().contains("User: first does not have permission='MANAGE' on address activemq.management"));
       }
 
       // BROWSE
@@ -641,8 +652,7 @@ public class SecurityTest extends ActiveMQTestBase {
          ClientConsumer browser = session.createConsumer(DURABLE_QUEUE, true);
          fail("should throw exception here");
       } catch (ActiveMQException e) {
-         assertTrue(e.getMessage().contains("User: first"));
-         assertTrue(e.getMessage().contains("does not have permission='BROWSE' for queue durableQueue on address address"));
+         assertTrue(e.getMessage().contains("User: first does not have permission='BROWSE' for queue durableQueue on address " + ADDRESS));
       }
    }
 
@@ -1041,7 +1051,16 @@ public class SecurityTest extends ActiveMQTestBase {
 
    @Test
    public void testJAASSecurityManagerAuthorizationPositive() throws Exception {
-      final SimpleString ADDRESS = SimpleString.of("address");
+      internalTestJAASSecurityManagerAuthorizationPositive(false);
+   }
+
+   @Test
+   public void testJAASSecurityManagerAuthorizationPositiveWithUuidNamespace() throws Exception {
+      internalTestJAASSecurityManagerAuthorizationPositive(true);
+   }
+
+   private void internalTestJAASSecurityManagerAuthorizationPositive(boolean useUuidNamespace) throws Exception {
+      final SimpleString ADDRESS = useUuidNamespace ? SimpleString.of(RandomUtil.randomUUIDString()) : SimpleString.of("address");
       final SimpleString DURABLE_QUEUE = SimpleString.of("durableQueue");
       final SimpleString NON_DURABLE_QUEUE = SimpleString.of("nonDurableQueue");
 
@@ -1049,24 +1068,34 @@ public class SecurityTest extends ActiveMQTestBase {
       ActiveMQServer server = addServer(ActiveMQServers.newActiveMQServer(createDefaultInVMConfig().setSecurityEnabled(true), ManagementFactory.getPlatformMBeanServer(), securityManager, false));
       Set<Role> roles = new HashSet<>();
       roles.add(new Role("programmers", true, true, true, true, true, true, true, true, true, true, false, false));
-      server.getConfiguration().putSecurityRoles("#", roles);
+      String match = "#";
+      if (useUuidNamespace) {
+         final SimpleString UUID_NAMESPACE = RandomUtil.randomUUIDSimpleString();
+         server.getConfiguration().setUuidNamespace(UUID_NAMESPACE.toString());
+         match = UUID_NAMESPACE + ".#";
+      }
+      server.getConfiguration().putSecurityRoles(match, roles);
       server.start();
 
       ClientSessionFactory cf = createSessionFactory(locator);
       ClientSession session = addClientSession(cf.createSession("first", "secret", false, true, true, false, 0));
 
       // CREATE_DURABLE_QUEUE
-      try {
-         session.createQueue(QueueConfiguration.of(DURABLE_QUEUE).setAddress(ADDRESS));
-      } catch (ActiveMQException e) {
-         fail("should not throw exception here");
+      if (!useUuidNamespace) {
+         try {
+            session.createQueue(QueueConfiguration.of(DURABLE_QUEUE).setAddress(ADDRESS));
+         } catch (ActiveMQException e) {
+            fail("should not throw exception here");
+         }
       }
 
       // DELETE_DURABLE_QUEUE
-      try {
-         session.deleteQueue(DURABLE_QUEUE);
-      } catch (ActiveMQException e) {
-         fail("should not throw exception here");
+      if (!useUuidNamespace) {
+         try {
+            session.deleteQueue(DURABLE_QUEUE);
+         } catch (ActiveMQException e) {
+            fail("should not throw exception here");
+         }
       }
 
       // CREATE_NON_DURABLE_QUEUE
@@ -1101,11 +1130,13 @@ public class SecurityTest extends ActiveMQTestBase {
       }
 
       // MANAGE
-      try {
-         ClientProducer producer = session.createProducer(server.getConfiguration().getManagementAddress());
-         producer.send(session.createMessage(true));
-      } catch (ActiveMQException e) {
-         fail("should not throw exception here");
+      if (!useUuidNamespace) {
+         try {
+            ClientProducer producer = session.createProducer(server.getConfiguration().getManagementAddress());
+            producer.send(session.createMessage(true));
+         } catch (ActiveMQException e) {
+            fail("should not throw exception here");
+         }
       }
 
       // BROWSE
@@ -1455,7 +1486,7 @@ public class SecurityTest extends ActiveMQTestBase {
    }
 
    @Test
-   public void testCreateTempQueueWithRole() throws Exception {
+   public void testCreateUuidWithRole() throws Exception {
       ActiveMQServer server = createServer();
 
       server.start();
@@ -1474,7 +1505,7 @@ public class SecurityTest extends ActiveMQTestBase {
    }
 
    @Test
-   public void testCreateTempQueueWithoutRole() throws Exception {
+   public void testCreateUuidWithoutRole() throws Exception {
       ActiveMQServer server = createServer();
 
       server.start();
@@ -1500,7 +1531,7 @@ public class SecurityTest extends ActiveMQTestBase {
    }
 
    @Test
-   public void testDeleteTempQueueWithRole() throws Exception {
+   public void testDeleteUuidWithRole() throws Exception {
       ActiveMQServer server = createServer();
       server.start();
       HierarchicalRepository<Set<Role>> securityRepository = server.getSecurityRepository();
@@ -1519,7 +1550,7 @@ public class SecurityTest extends ActiveMQTestBase {
    }
 
    @Test
-   public void testDeleteTempQueueWithoutRole() throws Exception {
+   public void testDeleteUuidWithoutRole() throws Exception {
       ActiveMQServer server = createServer();
       server.start();
       HierarchicalRepository<Set<Role>> securityRepository = server.getSecurityRepository();
